pyproject.toml
[tool.poetry]
name = "projetofinal-back"
version = "0.1.0"
description = ""
authors = ["Dimas_Marcondes <dimasmarcondes88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
django = "5.1.1"
django-cors-headers = "4.4.0"
django-extensions = "3.2.3"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
pillow = "10.4.0"
psycopg2 = "2.9.9"
pyjwt = "2.9.0"
sqlparse = "0.5.1"
tzdata = "2024.2"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
